+++ tiger-common/src/main/java/de/gematik/test/tiger/common/config/TigerConfigurationHelper.java
@@ -4,10 +4,9 @@
 
 package de.gematik.test.tiger.common.config;
 
-import java.util.Map;
+import de.gematik.test.tiger.common.util.TigerSerializationUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONObject;
-import org.yaml.snakeyaml.Yaml;
 
 @Slf4j
 @SuppressWarnings("unused")
@@ -22,8 +21,6 @@ public class TigerConfigurationHelper<T> {
    * @return JSON object representing the yaml content
    */
   public static JSONObject yamlStringToJson(String yamlStr) {
-    Yaml yaml = new Yaml(new DuplicateMapKeysForbiddenConstructor());
-    Map<String, Object> map = yaml.load(yamlStr);
-    return new JSONObject(map);
+    return TigerSerializationUtil.yamlToJsonObject(yamlStr);
   }
 }