+++ tiger-test-lib/src/test/java/de/gematik/test/tiger/lib/integrationtest/TestSaveLoadConfig.java
@@ -1,0 +1,18 @@
+/*
+ * ${GEMATIK_COPYRIGHT_STATEMENT}
+ */
+
+package de.gematik.test.tiger.lib.integrationtest;
+
+import io.cucumber.junit.CucumberOptions;
+import io.cucumber.junit.TigerCucumberRunner;
+import org.junit.runner.RunWith;
+
+@RunWith(TigerCucumberRunner.class)
+@CucumberOptions(
+    features = {
+      "./src/test/resources/features/tiger-test-lib/TigerGlue_SaveLoadGlobalConfig.feature"
+    },
+    glue = {"de.gematik.test.tiger.glue"},
+    tags = "not @Ignore")
+public class TestSaveLoadConfig {}